Code review refactor.
