--- src/components/base-scatterplot/index.js
@@ -1,2 +1,0 @@
-export { default as BaseScatterplotSubscriber } from './BaseScatterplotSubscriber';
-export { default as Scatterplot } from './Scatterplot';
