+++ src/components/embedding-scatterplot/index.js
@@ -1,0 +1,1 @@
+export { default as EmbeddingScatterplotSubscriber } from './EmbeddingScatterplotSubscriber';
