+++ src/components/scatterplot/ScatterplotOptions.js
