+++ src/components/scatterplot/ScatterplotTooltipSubscriber.js
