+++ src/components/scatterplot/index.js
@@ -1,1 +1,2 @@
 export { default as ScatterplotSubscriber } from './ScatterplotSubscriber';
+export { default as Scatterplot } from './Scatterplot';
